// pos-server/src/purchase-product/dto/purchase-report.dto.ts
import { CreatePurchaseProductDto } from "./create-purchase-product.dto";

export class PurchaseReport {

    invoiceNo: string;

    registerNo: string;

    supplierId: number;

    supplierName: string;
        
    date: Date;

    casherName: string;

    total: number;

    status: string;

    remark: string;

    purchaseItemsList: CreatePurchaseProductDto[]

}